Feat(FE): 회원탈퇴 확인 모달 추가
Former-commit-id: 54603e7c9ebc33c7fb5e4dd11c58accd72f44478

+++ frontend/src/views/MainPage.vue
@@ -275,6 +275,7 @@ export default {
   components: {
     LinkShareModal,
   },
+
   methods: {
     ...mapMutations(presentStore,["RESET_PRESENT_LIST"]),
     goToMainPage() {
@@ -491,7 +492,6 @@ export default {
         confirmButtonText: '수정',
         denyButtonText: `회원 탈퇴`,
         preConfirm: () => {
-          //validation 설정
           var exptext = /^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$/i;
           if (!flatpickrInstance.selectedDates[0]) {
             this.$swal.showValidationMessage(`생일을 입력해주세요`);
@@ -525,15 +525,24 @@ export default {
               res.data.User.userBirth[2];
           });
         } else if (result.isDenied) {
-          deleteUser(() => {
-            this.$swal.fire('회원 탈퇴되었습니다.', '', 'success')
-            localStorage.removeItem("accessToken")
-            this.$router.push({name:"Landing"})
-          })
+          this.$swal.fire({
+              title: '정말 회원을 탈퇴하시겠습니까?',
+              showCancelButton: true,
+            }).then((result) => {
+              if (result.isConfirmed) {
+                deleteUser(() => {
+                  this.$swal.fire('회원 탈퇴되었습니다.', '', 'success')
+                  localStorage.removeItem("accessToken")
+                  this.$store.commit("userStore/setUSerNull")
+                  this.$router.push({name:"Landing"})
+                })
+              }
+            })
         }
       })
     },
   },
+
 
 };
 </script>
